+++ CotEditor/Sources/EditorTextView+LineProcessing.swift
@@ -196,7 +196,7 @@ extension EditorTextView {
     /// show pattern sort sheet
     @IBAction func patternSort(_ sender: Any?) {
         
-        let viewController = NSStoryboard(name: NSStoryboard.Name("PatternSortView"), bundle: nil).instantiateInitialController() as! PatternSortViewController
+        let viewController = NSStoryboard(name: "PatternSortView", bundle: nil).instantiateInitialController() as! PatternSortViewController
         viewController.representedObject = self
         
         // sample the first line
@@ -207,7 +207,7 @@ extension EditorTextView {
         viewController.sampleLine = String(string[lineRange])
         viewController.sampleFontName = (self.layoutManager as? LayoutManager)?.textFont?.fontName
         
-        self.viewControllerForSheet?.presentViewControllerAsSheet(viewController)
+        self.viewControllerForSheet?.presentAsSheet(viewController)
     }