+++ CotEditor/Sources/EditorTextView+TouchBar.swift
@@ -25,14 +25,12 @@
 
 import Cocoa
 
-@available(macOS 10.12.2, *)
 private extension NSTouchBar.CustomizationIdentifier {
     
     static let textView = NSTouchBar.CustomizationIdentifier("com.coteditor.CotEditor.touchBar.textView")
 }
 
 
-@available(macOS 10.12.2, *)
 extension NSTouchBarItem.Identifier {
     
     static let shift = NSTouchBarItem.Identifier("com.coteditor.CotEditor.TouchBarItem.shift")
@@ -42,7 +40,6 @@ extension NSTouchBarItem.Identifier {
 
 
 
-@available(macOS 10.12.2, *)
 extension EditorTextView {
     
     override func makeTouchBar() -> NSTouchBar? {
@@ -92,7 +89,6 @@ extension EditorTextView {
 
 
 
-@available(macOS 10.12.2, *)
 extension EditorTextView {
     
     // MARK: NSCandidateListTouchBarItemDelegate
@@ -115,7 +111,6 @@ extension EditorTextView {
 }
     
 
-@available(macOS 10.12.2, *)
 extension EditorTextViewController {
     
     // MARK: NSTextViewDelegate