+++ CotEditor/Sources/EditorTextViewController.swift
@@ -50,11 +50,6 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
         
         set {
             self.scrollView?.rulersVisible = newValue
-            
-            // -> Workaround issue on Mojave where line number view covers text view.
-            if #available(macOS 10.14, *) {
-                self.scrollView?.layoutSubtreeIfNeeded()
-            }
         }
     }
     
@@ -199,7 +194,7 @@ final class EditorTextViewController: NSViewController, NSTextViewDelegate {
                 return
         }
         
-        self.presentViewControllerAsSheet(viewController)
+        self.presentAsSheet(viewController)
     }