+++ CotEditor/Sources/FindPanelFieldViewController.swift
@@ -33,8 +33,8 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate, 
     
     private weak var currentResultMessageTarget: NSLayoutManager?  // grab layoutManager instead of NSTextView to use weak reference
     
-    private lazy var regexReferenceViewController = DetachablePopoverViewController(nibName: NSNib.Name("RegexReferenceView"), bundle: nil)
-    private lazy var preferencesViewController = NSViewController(nibName: NSNib.Name("FindPreferencesView"), bundle: nil)
+    private lazy var regexReferenceViewController = DetachablePopoverViewController(nibName: "RegexReferenceView", bundle: nil)
+    private lazy var preferencesViewController = NSViewController(nibName: "FindPreferencesView", bundle: nil)
     
     @IBOutlet private var findTextView: RegexFindPanelTextView?  // NSTextView cannot be weak
     @IBOutlet private var replacementTextView: RegexFindPanelTextView?  // NSTextView cannot be weak
@@ -177,12 +177,12 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate, 
     @IBAction func showRegexHelp(_ sender: Any?) {
         
         if self.presentedViewControllers?.contains(self.regexReferenceViewController) ?? false {
-            self.dismissViewController(self.regexReferenceViewController)
+            self.dismiss(self.regexReferenceViewController)
             
         } else {
             guard let senderView = sender as? NSView else { return }
             
-            self.presentViewController(self.regexReferenceViewController, asPopoverRelativeTo: senderView.bounds, of: senderView, preferredEdge: .maxY, behavior: .semitransient)
+            self.present(self.regexReferenceViewController, asPopoverRelativeTo: senderView.bounds, of: senderView, preferredEdge: .maxY, behavior: .semitransient)
         }
     }
     
@@ -191,12 +191,12 @@ final class FindPanelFieldViewController: NSViewController, NSTextViewDelegate, 
     @IBAction func showPreferences(_ sender: Any?) {
         
         if self.presentedViewControllers?.contains(self.preferencesViewController) ?? false {
-            self.dismissViewController(self.preferencesViewController)
+            self.dismiss(self.preferencesViewController)
             
         } else {
             guard let senderView = sender as? NSView else { return }
             
-            self.presentViewController(self.preferencesViewController, asPopoverRelativeTo: senderView.bounds, of: senderView, preferredEdge: .maxX, behavior: .transient)
+            self.present(self.preferencesViewController, asPopoverRelativeTo: senderView.bounds, of: senderView, preferredEdge: .maxX, behavior: .transient)
         }
     }